+++ src/lib/ast/features/utils/index.ts
@@ -1,2 +1,2 @@
 
-export * from './math';
+export * from '../../math/math';